Free cells with expired pheromones are now considered as UNKNOWN, also fixed some visualization issues

+++ argos3-examples/loop_functions/agent_vision_loop_functions_pipuck/agent_vision_loop_functions_pipuck.h
@@ -16,8 +16,8 @@ public:
     TCoordinateMap m_tOtherAgentCoordinates;
     std::map<CPiPuckEntity*, CVector3> m_tAgentCoordinates;
     std::map<CPiPuckEntity*, CVector3> m_tAgentBestFrontierCoordinate;
-    std::map<CPiPuckEntity*, std::vector<std::tuple<quadtree::Box, int, uint32_t >>> m_tQuadTree;
-    std::map<CPiPuckEntity*, uint32_t> m_tAgentElapsedTicks;
+    std::map<CPiPuckEntity*, std::vector<std::tuple<quadtree::Box, int, double >>> m_tQuadTree;
+    std::map<CPiPuckEntity*, double> m_tAgentElapsedTicks;
 
 public:
 
@@ -41,11 +41,11 @@ public:
         return m_tAgentCoordinates;
     }
 
-    inline const std::map<CPiPuckEntity*, std::vector<std::tuple<quadtree::Box, int, uint32_t >>>& GetQuadTree() const {
+    inline const std::map<CPiPuckEntity*, std::vector<std::tuple<quadtree::Box, int, double >>>& GetQuadTree() const {
         return m_tQuadTree;
     }
 
-    inline const std::map<CPiPuckEntity*, uint32_t>& GetAgentElapsedTicks() const {
+    inline const std::map<CPiPuckEntity*, double>& GetAgentElapsedTicks() const {
         return m_tAgentElapsedTicks;
     }
